Remove excess file
